requirements.txt
pdfplumber
PyPDF2
ttkbootstrap
pytesseract
Pillow
python-dateutil
requests
requests
